requirements.txt
opencv-contrib-python>=4.7.0
numpy>=1.24.0
pillow>=9.0.0